+++ src/egeo-demo/st-form-demo/json-schema.ts
@@ -11,54 +11,74 @@
 export const JSON_SCHEMA: any = {
    '$schema': 'http://json-schema.org/schema#',
    'properties': {
-      'genericNumberInput': {
+      'genericIntegerInput': {
          'title': 'Generic number',
          'description': 'Generic input description',
-         'type': 'number',
+         'type': 'integer',
          'default': 5,
          'minimum': 6,
          'maximum': 10,
          'exclusiveMinimum': false,
-         'exclusiveMaximum': true
+         'exclusiveMaximum': false
+      },
+      'genericNumberInput': {
+         'title': 'Generic number',
+         'description': 'Generic input description',
+         'type': 'number',
+         'default': 1,
+         'minimum': 0.5,
+         'maximum': 10,
+         'exclusiveMinimum': false,
+         'exclusiveMaximum': false
       },
       'requiredNumber': {
          'title': 'Required number',
          'description': 'Required input description',
-         'type': 'number',
+         'type': 'integer',
          'default': 5
       },
       'minNumber': {
          'title': 'Min number',
          'description': 'Min number input description',
          'type': 'number',
-         'default': 28017,
+         'default': 8.6,
          'minimum': 6
       },
       'maxNumber': {
          'title': 'Max number',
          'description': 'Max number input description',
          'type': 'number',
-         'default': 28017,
-         'maximum': 5
+         'default': 7.6,
+         'maximum': 6
       },
       'minAndMaxNumber': {
          'title': 'Number in a range',
          'description': 'This number has to be between 7 and 19',
          'type': 'number',
-         'default': 28017,
+         'default': 8.6,
          'minimum': 6,
-         'maximum': 20,
+         'maximum': 10,
          'exclusiveMinimum': true,
          'exclusiveMaximum': true
+      },
+
+      'genericTextInput': {
+         'title': 'Required text',
+         'description': 'This is a required text with a length of 2-6 characters and only the character a is valid',
+         'type': 'string',
+         'default': 'Generic text',
+         'minLength': 2,
+         'maxLength': 6,
+         'pattern': '(a)+'
       },
       'requiredText': {
          'title': 'Required text',
          'description': 'This is a required text',
          'type': 'string',
-         'default': 'required text'
+         'default': '5'
       },
       'minLengthText': {
-         'title': 'Text with a min length',
+         'title': 'Text with a max length',
          'description': 'You have to type a text with 10 characters at least',
          'type': 'string',
          'default': '',
@@ -69,8 +89,7 @@ export const JSON_SCHEMA: any = {
          'description': 'You have to type a text with less than 20 characters',
          'type': 'string',
          'default': '',
-         'maxLength': 20,
-         'pattern': 'aa+'
+         'maxLength': 20
       },
       'minAndMaxLengthText': {
          'title': 'Text with a min and max length',
@@ -78,8 +97,7 @@ export const JSON_SCHEMA: any = {
          'type': 'string',
          'default': '',
          'minLength': 10,
-         'maxLength': 20,
-
+         'maxLength': 20
       },
       'url': {
          'title': 'URL',
@@ -92,9 +110,13 @@ export const JSON_SCHEMA: any = {
          'description': 'You can enable or disable this property',
          'type': 'boolean',
          'default': true
+      },
+      'noDescription': {
+         'title': 'Property without description',
+         'type': 'string'
       }
    },
    'required': [
-      'requiredNumber', 'requiredText', 'url'
+      'url', 'genericNumberInput', 'genericIntegerInput', 'requiredNumber', 'requiredText'
    ]
 };